+++ galsim/utilities.py
@@ -739,7 +739,7 @@ def deInterleaveImage(image, N, conserve_flux=False,suppress_warnings=False):
                 img.wcs = img_wcs
             ## Preserve the origin so that the interleaved image has the same bounds as the image
             ## that is being deinterleaved.
-            img.setOrigin(image.origin())
+            img.setOrigin(image.origin)
 
     elif suppress_warnings is False:
         import warnings
@@ -913,10 +913,10 @@ def interleaveImages(im_list, N, offsets, add_flux=True, suppress_warnings=False
         warnings.warn("Interleaved image could not be assigned a WCS automatically.")
 
     # Assign a possibly non-trivial origin and warn if individual image have different origins.
-    orig = im_list[0].origin()
+    orig = im_list[0].origin
     img.setOrigin(orig)
     for im in im_list[1:]:
-        if not im.origin()==orig:  # pragma: no cover
+        if not im.origin==orig:  # pragma: no cover
             import warnings
             warnings.warn("Images in `im_list' have multiple values for origin. Assigning the \
             origin of the first Image instance in 'im_list' to the interleaved image.")
@@ -1268,7 +1268,7 @@ def unweighted_moments(image, origin=galsim.PositionD(0, 0)):
     @returns  Dict with entries for [M0, Mx, My, Mxx, Myy, Mxy]
     """
     a = image.array.astype(float)
-    offset = image.origin() - origin
+    offset = image.origin - origin
     xgrid, ygrid = np.meshgrid(np.arange(image.array.shape[1]) + offset.x,
                                np.arange(image.array.shape[0]) + offset.y)
     M0 = np.sum(a)