+++ galsim/wcs.py
@@ -503,11 +503,11 @@ class BaseWCS(object):
         So, for example, to set a WCS that has a constant pixel size with the world coordinates
         centered at the center of an image, you could write:
 
-            >>> wcs = galsim.PixelScale(scale).withOrigin(im.center())
+            >>> wcs = galsim.PixelScale(scale).withOrigin(im.center)
 
         This is equivalent to the following:
 
-            >>> wcs = galsim.OffsetWCS(scale, origin=im.center())
+            >>> wcs = galsim.OffsetWCS(scale, origin=im.center)
 
         For non-local WCS types, the origin defines the location in the image coordinate system
         should mean the same thing as (x,y) = (0,0) does for the current WCS.  The following
@@ -2240,7 +2240,7 @@ class UVFunction(EuclideanWCS):
         _writeFuncToHeader(self._orig_xfunc, 'X', header)
         _writeFuncToHeader(self._orig_yfunc, 'Y', header)
 
-        return self.affine(bounds.trueCenter())._writeLinearWCS(header, bounds)
+        return self.affine(bounds.true_center)._writeLinearWCS(header, bounds)
 
     @staticmethod
     def _readHeader(header):
@@ -2396,7 +2396,7 @@ class RaDecFunction(CelestialWCS):
         _writeFuncToHeader(self._orig_ra_func, 'R', header)
         _writeFuncToHeader(self._orig_dec_func, 'D', header)
 
-        return self.affine(bounds.trueCenter())._writeLinearWCS(header, bounds)
+        return self.affine(bounds.true_center)._writeLinearWCS(header, bounds)
 
     @staticmethod
     def _readHeader(header):